requirements.txt
# Required packages for ComfyUI Ideogram Character Node
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
Pillow>=10.0.0
requests>=2.28.0